requirements.txt
Django>=5.0,<6.0
djangorestframework==3.15.2
Pillow>=10.0.0
python-dotenv==1.0.1
anthropic>=0.28.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-cors-headers==4.4.0
whitenoise==6.7.0
gunicorn==22.0.0
Markdown==3.6
